Lab8/requirements.txt
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
future==0.18.2
iso8601==1.1.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
